New Data pull, multiple year reports
Updating the test data to be my latest pull. I also modified the main
method to print reports for an array of years. Eventually I want to
modify the yearly report service to take a set of years instead to not
only do reports for each year, but also year to year statistics. For now
this is good enough. 

Maybe next year. :)