First Pass at Gender Report
I created a new enricher that will set the author's gender based on how
the book is shelved.

I re-factored some of the Genre stuff in the process, but more
re-factoring is needed. However, it requires redoing the yearly report
stuff first and then consolidating more code into the abstract report
service.

Added some additional documentation to the ReadME.

Updated the data file.