Changed YearlyReportService to take in one or more years.
The original implementation only allowed for a single year requiring you
to create new service for every year you wanted a report for. Now it
takes in an array (using Integer...) so you can have 1 or more years.

This not only simplifies the calling code but will allow for
year-to-year comparisons as well.