Adding Year-to-Year Gender report
I think if I'm unable to determine the gender based on shelves
(typically this is for books with multiple authors/contributors) I'm not
setting a gender at all, so those numbers are being excluded and the
percentages don't add up to (around) 100% like other reports do.

I'm not sure if defaulting to Unknown gender would be useful though, so
I'm leaving it as is for now.