Made some tweaks to the yearly report
1) If the Original Year of Publication is missing/null it will first try
the edition's year of publication before logging a warning/defaulting to
the year of the report

2) Graphic Novels will no longer be compared for longest/shortest books.
They'll still be counted for the number of pages read however. Basically
I found that every year my shortest book was always a graphic novel
which was uninteresting to me.

But I already break out the pages read to exclude graphic novels
allowing me to see total both with and without them.

3) Added a header for the Book Ratings to the report output.